requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
anthropic==0.39.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
aiofiles==24.1.0
jinja2==3.1.4
